Supporting even larger timestamps
We now format the whole part and the decimal part separately.

Since the whole part is by definition, we need not covert it to a float,
hence the whole part can simply be formatted as an integer.

The and since by definition of the `Time.Nanoseconds()`, the return
value is between 0-999999999, only this part needs to be formatted
separately.

We then combine whole + decimals[1:] to form the final result.

This allows us to correctly format the maximun timestamp that is allowed
by Go.