Bump Go version to indicate correct minimum requirement
The minimum requirement that we have was silently bumped to 1.21 in https://github.com/golang-jwt/jwt/pull/441 because of the `slices` package. It seems that we did not update the `go.mod` when we updated our CI range, because CI did not fail because it was not testing older versions.

We probably should just update the `go.mod` when we update the CI target in the future? Although we could theoretically stay at 1.21 in terms of the code base.